Atualização no cadastro de gestor e palestrante , agora são gerados atm

+++ src/main/java/com/ifpe/semanainfo/repository/Speaks.java
@@ -8,9 +8,10 @@ import com.ifpe.semanainfo.model.Speaker;
 
 
 
-public interface Speaks extends JpaRepository<Speaker, Long> {
+public interface Speaks extends JpaRepository<Speaker, Integer> {
 	
 	public Optional<Speaker> findByEmail(String email);
+
 	
 
 }
